\name{print}
\alias{print.tscausal}
\title{
Prints the output of the causal analysis
}
\description{
Prints a summary table of the causal analysis.
}
\usage{
\method{print}{tscausal}(x, digits = 4, ...)
}
\arguments{
\item{x}{
An object of class \dQuote{tscausal} generated from
calling \code{\link{tscausal}}.
}
\item{digits}{
Minimal number of significant digits.
}
\item{\dots}{
Not used.
}
}
\author{
Alexios Galanos
}